ndt-server.go: integrate with bassosimone/nuvolari
The bassosimone/nuvolari contains a prototype implementation of
a next generation NDT where we remove the requirement to be backward
compatible and embrace WebSocket and TLS to the maximum.

At the moment of writing this commit message, bassosimone/nuvolari
has not been pushed on GitHub yet but I'll remediate soon.