BBR: avoid possibly confusing measurement units
In the kernel, the data is represented as such because integer
arithmetic is used and precision is important.

Since we use float64, we may as well use the most natural units
with which people usually reasons about these variables.

Speed
`````

For the speed, the correct value is bit/s, because ISPs sell
their connections using that measurement unit. So, one expects
that measurement unit when reasoning about his/her speed.

I have observed what speedtest.net and fast.com do. The former
apparently is using always Mbit/s, while the latter scales the
number to the most appropriate unit (e.g. Kbit/s). By using
bit/s by default, we enable applications to choose what to do
according to the developers taste easily.

The important thing is to avoid byte/s, because that is quite
unusual and hence may really lead to confusion, if the developer
integrating NDTv7 is not paying enough attention.

RTT
```

For the RTT, the correct value is perhaps milliseconds, which
is the unit used by ping and speedtest.

There MAY also be an argument in favour of seconds, but really it
seems to me the unit the bears with it the least amount of possible
developer errors is actually milliseconds.

So, milliseconds.