Add ndt7 spec and sync the implementation
Most likely always using float64 is going to save us headaches
in the event in which we're dealing with 32 bit clients.

There are still 32 bit devices around (i.e. Android and iOS devices)
so I believe that this is a valid concern.

Incidentally, this move remove any overflow concern when adding
up bytes because there is no overflow with float64 just a loss of
precision when you use more than 53 bits for the value.