Use go1.11 on docker and statically link binary
The original code used CGO_ENABLED=0 to make sure that the net
implementation did not link on libc, hence later it was possible
to use a very thin distribution such as Alpine, which uses
musl libc. Now, with bbr, we cannot avoid compiling using cgo,
so we need another strategy to disable using glibc.

That is using the `-tags netgo` tag to build. The resulting
binary is statically linked for me, and runs on the destination
Alpine container. I cannot access it, however, because Docker
for Mac works differently (well, maybe with more effort I could
manage to make it work, but after 15' I could not).

References:

[1] https://www.reddit.com/r/golang/comments/8m4xrh/do_linux_golang_binaries_depend_on_libc/

[2] https://forums.docker.com/t/should-docker-run-net-host-work/14215/28