bbr: explicitly flag _max_ bandwidth and _min_ rtt
While that was already clear in the documentation, it was not
explicit in variable names and in JSON message fields.

Yesterday, talking to @mattmathis, I made the mistake of
considering the min-RTT being the RTT, which led me to draw
wrong conclusions regarding a connection.

For this reason, I have decided to go a bit more explicit
with variable names. If the person who wrote the code, i.e.
me, wrongly interprets variables, this may happen as well
to someone less familiar with the codebase. Hence, it is
clearly better to have a much less ambiguous naming.

No need to bump the specification version because it has
already been bumped as part of this PR.