Move MinMeasurementInterval away from spec
Since version 0.3.0 of the spec, there is no min-measurement-interval
as it's possible to use just measurement messages to load the net.

Hence, just move the variable inside the client.go file.

Sometime in the future, client.go will be moved in another place
and become a unit test for the server.