ndt7: start refactoring to reduce the number of channels (#248)
As documented in #238, we would like to reduce the pipeline
structure and the dependency on channels. We have a reasonably
clear idea of what we want to do.

Let us do this piecemeal.

In this first diff, I'm changing the measurer to be a
struct rather than just a func. By doing that, I'm
creating the opportunity for storing the measurements
collected by the measurer into the measurer itself rather
than publishing them downstream to a channel.